Moving JwtBuilder to Builder folder
